Rooms and Fixes
I began to add some of the room nodes of the world map. However, I have
not added their descriptions yet, since the story is still being
planned. Also, aside from that, I fixed an issue in which the program
did not let you to quiet when you typed the specified commands.